implement project and split informations
