+++ src/app/config/contact-config.ts
@@ -1,0 +1,8 @@
+// src/app/config/contact-config.ts
+
+export const ContactConfig = {
+    email: 'max@example.com',
+    twitter: 'maxmustermann',
+    linkedin: 'maxmustermann',
+    github: 'max-mustermann'
+  };
