+++ src/app/config/footer-config.ts
@@ -1,0 +1,11 @@
+// src/app/config/footer-config.ts
+
+export const FooterConfig = {
+    tagline: 'Full-Stack Developer specializing in modern web technologies.',
+    labels: {
+      pages: 'Pages',
+      connect: 'Connect'
+    },
+    copyrightName: 'Max Mustermann',
+    credits: 'Made with ❤ as an open-source template'
+  };
