+++ src/app/config/header-config.ts
@@ -1,0 +1,16 @@
+// src/app/config/header-config.ts
+
+export const HeaderConfig = {
+    logoImage: {
+      path: 'assets/images/logo_dark.png',  // Path to your logo
+      alt: 'Portfolio Logo',
+      width: 200,
+      height: 80
+    },
+    navigation: {
+      home: 'Home',
+      projects: 'Projects', 
+      about: 'About',
+      skills: 'Skills'
+    }
+  };
